Minor changes in processing simulation events

+++ Fitter_if.h
@@ -23,7 +23,7 @@ public:
 	virtual void fitTriangular(double *sqrerror, double *min, double *mo, double *max) = 0;
 	virtual void fitNormal(double *sqrerror, double *avg, double *stddev) = 0;
 	virtual void fitExpo(double *sqrerror, double *avg1) = 0;
-	virtual void fitErlang(double *sqrerror, double *avg, int *m) = 0;
+	virtual void fitErlang(double *sqrerror, double *avg, double *m) = 0;
 	virtual void fitBeta(double *sqrerror, double *alpha, double *beta, double *infLimit, double *supLimit) = 0;
 	virtual void fitWeibull (double *sqrerror, double *alpha, double *scale)  = 0;
 	virtual void fitAll(double *sqrerror, std::string *name) = 0;
